Print the first 50 Fibonacci numbers, starting with 1 and 2, followed by a new line

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -19,13 +19,13 @@
 int main(void)
 {
 	int i, n = 50;
-	unsigned long long int  fib1 = 1, fib2 = 2, next;
+	unsigned long int  fib1 = 1, fib2 = 2, next;
 
-	printf("%llu, %llu", fib1, fib2);
+	printf("%lu, %lu", fib1, fib2);
 	for (i = 2; i < n; i++)
 	{
 		next = fib1 + fib2;
-		printf(", %llu", next);
+		printf(", %lu", next);
 		fib1 = fib2;
 		fib2 = next;
 	}
